updated all files

+++ src/main/java/idespring/lab2/repository/studentrepo/StudentRepositoryCustom.java
@@ -4,7 +4,7 @@ import idespring.lab2.model.Student;
 import java.util.List;
 
 public interface StudentRepositoryCustom {
-    List<Student> findByAgeAndSortByName(long age, String sort);
+    List<Student> findByAgeAndSortByName(int age, String sort);
 
     List<Student> sortByName(String sort);
 }
